requirements.txt
opencv-python>=4.8.0
numpy>=1.21.0
insightface>=0.7.3
PyQt5>=5.15.0

# GPU Acceleration (use onnxruntime-gpu for NVIDIA GPUs)
onnxruntime-gpu==1.20.0

# PyTorch for GPU acceleration (CUDA 12.6)
# Install separately:
# pip3 install torch torchvision --index-url https://download.pytorch.org/whl/cu126
